requirements.txt
streamlit
pandas
pyarrow
jsonschema
google-generativeai